Minor refractoring

+++ ImageManipulation/include/ImageManipulation/image.h
@@ -5,7 +5,7 @@ typedef char* Pixel;
 
 typedef struct Image {
 	unsigned int width, height;
-	unsigned int pixel_size;
+	unsigned short pixel_size;
 
 	char* data;
 
@@ -13,7 +13,7 @@ typedef struct Image {
 	char* header;
 } Image;
 
-Image* alloc_image(const unsigned int width, const unsigned int height, const unsigned int pixel_size, const unsigned int header_size);
+Image* alloc_image(const unsigned int width, const unsigned int height, const unsigned short pixel_size, const unsigned int header_size);
 void dealloc_image(Image* image);
 
 Image* duplicate_image(const Image* image);
